refactor: move members from old utils
