+++ src/electron/background.ts
@@ -2,6 +2,7 @@ import { app } from "electron";
 
 import { main } from "@/main";
 
+import "@/core-setup";
 import "@/ipc/listeners";
 
 main().then(() => {
