+++ src/shared/flags.ts
@@ -1,1 +1,3 @@
 export const isProduction = process.env.NODE_ENV === "production";
+export const isDevelopment =
+	process.env.NODE_ENV === "development" || process.env.NODE_ENV === "test";
